requirements.txt
pyyaml
paramiko
rich
netmiko